+++ backend/src/main/java/com/utsav/arts/dtos/artworkDTO/ArtworkResponseDTO.java
@@ -1,5 +1,7 @@
 package com.utsav.arts.dtos.artworkDTO;
 
+import java.math.BigDecimal;
+
 public class ArtworkResponseDTO {
 
     private int id;
@@ -7,7 +9,7 @@ public class ArtworkResponseDTO {
     private String description;
     private String category;
     private String label;
-    private int price;
+    private BigDecimal price;
     private String imgUrl;
 
     public ArtworkResponseDTO() {}
@@ -27,8 +29,8 @@ public class ArtworkResponseDTO {
     public String getLabel() { return label; }
     public void setLabel(String label) { this.label = label; }
 
-    public int getPrice() { return price; }
-    public void setPrice(int price) { this.price = price; }
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
 
     public String getImgUrl() { return imgUrl; }
     public void setImgUrl(String imgUrl) { this.imgUrl = imgUrl; }
